@@ -1,29 +1,49 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdkVersion 26
-    buildToolsVersion '27.0.3'
+    compileSdkVersion 33
     defaultConfig {
         applicationId "com.google.android.gms.example.nativeadvancedrecyclerviewexample"
-        minSdkVersion 14
-        targetSdkVersion 26
+        minSdkVersion 27
+        targetSdkVersion 33
         versionCode 1
         versionName "1.0"
+        multiDexEnabled true
     }
     buildTypes {
         release {
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
+        debug {
+            minifyEnabled false
+        }
     }
 }
 
 dependencies {
-    implementation 'com.google.firebase:firebase-ads:15.0.0'
-    implementation 'com.android.support:appcompat-v7:26.1.0'
-    implementation 'com.android.support:cardview-v7:26.1.0'
-    implementation 'com.android.support:recyclerview-v7:26.1.0'
-    implementation 'com.android.support:support-v4:26.1.0'
+    implementation "androidx.multidex:multidex:2.0.1"
+//    implementation 'com.google.firebase:firebase-ads:21.3.0'
+    implementation 'com.google.android.gms:play-services-ads:21.3.0'
+    implementation 'androidx.appcompat:appcompat:1.5.1'
+    implementation 'androidx.cardview:cardview:1.0.0'
+    implementation 'androidx.recyclerview:recyclerview:1.2.1'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation 'io.reactivex.rxjava2:rxandroid:2.1.1'
+    implementation 'io.reactivex.rxjava2:rxjava:2.2.19'
+    implementation 'com.jakewharton.timber:timber:5.0.1'
+
+    //paging
+    implementation 'androidx.paging:paging-runtime-ktx:3.1.1'
+    testImplementation 'androidx.paging:paging-common-ktx:3.1.1' // For Kotlin use paging-common-ktx
+
 }
+configurations {
+    all {
+        exclude group: 'androidx.lifecycle', module: 'lifecycle-runtime-ktx'
+        exclude group: 'androidx.lifecycle', module: 'lifecycle-viewmodel-ktx'
+    }
+}
+
 
 apply plugin: 'com.google.gms.google-services'
