@@ -16,7 +16,7 @@
 package com.google.android.gms.example.nativeadvancedrecyclerviewexample;
 
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
